fix(setcasecompletions, dynamodb): refactor updatecase to dynamodb module
